moved test to support
