+++ support/pom.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>no.dcat</groupId>
+        <artifactId>fdk</artifactId>
+        <version>0.2.4-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>support</artifactId>
+    <packaging>pom</packaging>
+
+
+    <name>Various support applications</name>
+
+
+    <modules>
+        <module>test</module>
+        <module>gdoc-import</module>
+    </modules>
+
+
+</project>
