+++ support/test/copyLibs.bat
