+++ support/test/src/main/docker/Dockerfile
