+++ support/test/src/main/java/no/dcat/testadmin/SecurityConfig.java
