+++ support/test/stop.cmd
